chore(hover-card): align file names
