+++ packages/primitives/hover-card/README.md
@@ -1,3 +1,3 @@
-# @radix-ng/primitives/tooltip
+# @radix-ng/primitives/hover-card
 
-Secondary entry point of `@radix-ng/primitives`. It can be used by importing from `@radix-ng/primitives/tooltip`.
+Secondary entry point of `@radix-ng/primitives`. It can be used by importing from `@radix-ng/primitives/hover-card`.
