+++ packages/primitives/hover-card/index.ts
@@ -1,19 +1,19 @@
 import { NgModule } from '@angular/core';
-import { RdxTooltipAnchorDirective } from './src/tooltip-anchor.directive';
-import { RdxTooltipArrowDirective } from './src/tooltip-arrow.directive';
-import { RdxTooltipCloseDirective } from './src/tooltip-close.directive';
-import { RdxTooltipContentAttributesComponent } from './src/tooltip-content-attributes.component';
-import { RdxTooltipContentDirective } from './src/tooltip-content.directive';
-import { RdxTooltipRootDirective } from './src/tooltip-root.directive';
-import { RdxTooltipTriggerDirective } from './src/tooltip-trigger.directive';
+import { RdxTooltipAnchorDirective } from './src/hover-card-anchor.directive';
+import { RdxTooltipArrowDirective } from './src/hover-card-arrow.directive';
+import { RdxTooltipCloseDirective } from './src/hover-card-close.directive';
+import { RdxTooltipContentAttributesComponent } from './src/hover-card-content-attributes.component';
+import { RdxTooltipContentDirective } from './src/hover-card-content.directive';
+import { RdxTooltipRootDirective } from './src/hover-card-root.directive';
+import { RdxTooltipTriggerDirective } from './src/hover-card-trigger.directive';
 
-export * from './src/tooltip-anchor.directive';
-export * from './src/tooltip-arrow.directive';
-export * from './src/tooltip-close.directive';
-export * from './src/tooltip-content-attributes.component';
-export * from './src/tooltip-content.directive';
-export * from './src/tooltip-root.directive';
-export * from './src/tooltip-trigger.directive';
+export * from './src/hover-card-anchor.directive';
+export * from './src/hover-card-arrow.directive';
+export * from './src/hover-card-close.directive';
+export * from './src/hover-card-content-attributes.component';
+export * from './src/hover-card-content.directive';
+export * from './src/hover-card-root.directive';
+export * from './src/hover-card-trigger.directive';
 
 const _imports = [
     RdxTooltipArrowDirective,
@@ -29,4 +29,4 @@ const _imports = [
     imports: [..._imports],
     exports: [..._imports]
 })
-export class RdxTooltipModule {}
+export class RdxHoverCardModule {}
