+++ packages/primitives/hover-card/src/hover-card-anchor.token.ts
@@ -1,4 +1,4 @@
 import { InjectionToken } from '@angular/core';
-import { RdxTooltipAnchorDirective } from './tooltip-anchor.directive';
+import { RdxTooltipAnchorDirective } from './hover-card-anchor.directive';
 
 export const RdxTooltipAnchorToken = new InjectionToken<RdxTooltipAnchorDirective>('RdxTooltipAnchorToken');
