+++ packages/primitives/hover-card/src/hover-card-arrow.token.ts
@@ -1,4 +1,4 @@
 import { InjectionToken } from '@angular/core';
-import { RdxTooltipArrowDirective } from './tooltip-arrow.directive';
+import { RdxTooltipArrowDirective } from './hover-card-arrow.directive';
 
 export const RdxTooltipArrowToken = new InjectionToken<RdxTooltipArrowDirective>('RdxTooltipArrowToken');
