+++ packages/primitives/hover-card/src/hover-card-close.token.ts
@@ -1,4 +1,4 @@
 import { InjectionToken } from '@angular/core';
-import { RdxTooltipCloseDirective } from './tooltip-close.directive';
+import { RdxTooltipCloseDirective } from './hover-card-close.directive';
 
 export const RdxTooltipCloseToken = new InjectionToken<RdxTooltipCloseDirective>('RdxTooltipCloseToken');
