+++ packages/primitives/hover-card/src/hover-card-content-attributes.token.ts
@@ -1,5 +1,5 @@
 import { InjectionToken } from '@angular/core';
-import { RdxTooltipContentAttributesComponent } from './tooltip-content-attributes.component';
+import { RdxTooltipContentAttributesComponent } from './hover-card-content-attributes.component';
 
 export const RdxTooltipContentAttributesToken = new InjectionToken<RdxTooltipContentAttributesComponent>(
     'RdxTooltipContentAttributesToken'
