+++ packages/primitives/hover-card/stories/hover-card-anchor.component.ts
@@ -1,8 +1,9 @@
-import { Component, signal, viewChild } from '@angular/core';
+import { Component, viewChild } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { LucideAngularModule, MountainSnow, TriangleAlert, X } from 'lucide-angular';
-import { RdxTooltipModule } from '../index';
-import { RdxTooltipContentAttributesComponent } from '../src/tooltip-content-attributes.component';
+import { LucideAngularModule, MapPin, MapPinPlus, MountainSnow, TriangleAlert, X } from 'lucide-angular';
+import { RdxHoverCardModule } from '../index';
+import { RdxTooltipAnchorDirective } from '../src/hover-card-anchor.directive';
+import { RdxTooltipContentAttributesComponent } from '../src/hover-card-content-attributes.component';
 import { provideRdxCdkEventService } from '../src/utils/cdk-event.service';
 import { containerAlert } from './utils/constants';
 import { OptionPanelBase } from './utils/option-panel-base.class';
@@ -10,18 +11,19 @@ import styles from './utils/styles.constants';
 import { WithOptionPanelComponent } from './utils/with-option-panel.component';
 
 @Component({
-    selector: 'rdx-tooltip-triggering',
+    selector: 'rdx-tooltip-anchor',
     providers: [provideRdxCdkEventService()],
     imports: [
         FormsModule,
-        RdxTooltipModule,
+        RdxHoverCardModule,
         LucideAngularModule,
         RdxTooltipContentAttributesComponent,
-        WithOptionPanelComponent
+        WithOptionPanelComponent,
+        RdxTooltipAnchorDirective
     ],
     styles: styles(),
     template: `
-        <p class="ExampleSubtitle">Initially closed</p>
+        <p class="ExampleSubtitle">Internal Anchor (within TooltipRoot)</p>
         <tooltip-with-option-panel
             [arrowWidth]="arrowWidth()"
             [arrowHeight]="arrowHeight()"
@@ -34,20 +36,6 @@ import { WithOptionPanelComponent } from './utils/with-option-panel.component';
             (openDelayChange)="openDelay.set($event)"
             (closeDelayChange)="closeDelay.set($event)"
         >
-            <div class="ParamsContainer">
-                <button (mouseup)="triggerOpenFalse()" type="button">Open: {{ isOpenFalse() }}</button>
-                onOpenChange count: {{ counterOpenFalse() }}
-            </div>
-
-            <div class="ParamsContainer">
-                <input
-                    [ngModel]="externalControlFalse()"
-                    (ngModelChange)="externalControlFalse.set($event)"
-                    type="checkbox"
-                />
-                External control
-            </div>
-
             <div class="ContainerAlerts">
                 <lucide-angular [img]="TriangleAlert" size="16" />
                 {{ containerAlert }}
@@ -55,22 +43,21 @@ import { WithOptionPanelComponent } from './utils/with-option-panel.component';
             <div class="container">
                 <ng-container
                     #root1="rdxTooltipRoot"
-                    [open]="isOpenFalse()"
                     [openDelay]="openDelay()"
                     [closeDelay]="closeDelay()"
-                    [externalControl]="externalControlFalse()"
                     rdxTooltipRoot
                 >
+                    <button class="reset IconButton InternalAnchor" rdxTooltipAnchor>
+                        <lucide-angular [img]="LucideMapPinPlusInside" size="16" style="display: flex" />
+                    </button>
+
                     <button class="reset IconButton" rdxTooltipTrigger>
                         <lucide-angular [img]="MountainSnowIcon" size="16" style="display: flex" />
                     </button>
 
                     <ng-template
-                        [sideOffset]="8"
                         [onOverlayEscapeKeyDownDisabled]="onOverlayEscapeKeyDownDisabled()"
                         [onOverlayOutsideClickDisabled]="onOverlayOutsideClickDisabled()"
-                        (onOpen)="countOpenFalse(true)"
-                        (onClosed)="countOpenFalse(false)"
                         rdxTooltipContent
                     >
                         <div class="TooltipContent" rdxTooltipContentAttributes>
@@ -91,7 +78,7 @@ import { WithOptionPanelComponent } from './utils/with-option-panel.component';
             <div class="TooltipId">ID: {{ rootDirective1()?.uniqueId() }}</div>
         </tooltip-with-option-panel>
 
-        <p class="ExampleSubtitle">Initially open</p>
+        <p class="ExampleSubtitle">External Anchor (outside TooltipRoot)</p>
         <tooltip-with-option-panel
             [arrowWidth]="arrowWidth()"
             [arrowHeight]="arrowHeight()"
@@ -104,31 +91,20 @@ import { WithOptionPanelComponent } from './utils/with-option-panel.component';
             (openDelayChange)="openDelay.set($event)"
             (closeDelayChange)="closeDelay.set($event)"
         >
-            <div class="ParamsContainer">
-                <button (mouseup)="triggerOpenTrue()" type="button">Open: {{ isOpenTrue() }}</button>
-                <span>onOpenChange count: {{ counterOpenTrue() }}</span>
-            </div>
-
-            <div class="ParamsContainer">
-                <input
-                    [ngModel]="externalControlTrue()"
-                    (ngModelChange)="externalControlTrue.set($event)"
-                    type="checkbox"
-                />
-                External control
-            </div>
-
             <div class="ContainerAlerts">
                 <lucide-angular [img]="TriangleAlert" size="16" />
                 {{ containerAlert }}
             </div>
             <div class="container">
+                <button class="reset IconButton ExternalAnchor" #externalAnchor="rdxTooltipAnchor" rdxTooltipAnchor>
+                    <lucide-angular [img]="LucideMapPinPlus" size="16" style="display: flex" />
+                </button>
+
                 <ng-container
                     #root2="rdxTooltipRoot"
-                    [open]="isOpenTrue()"
+                    [anchor]="externalAnchor"
                     [openDelay]="openDelay()"
                     [closeDelay]="closeDelay()"
-                    [externalControl]="externalControlTrue()"
                     rdxTooltipRoot
                 >
                     <button class="reset IconButton" rdxTooltipTrigger>
@@ -136,11 +112,8 @@ import { WithOptionPanelComponent } from './utils/with-option-panel.component';
                     </button>
 
                     <ng-template
-                        [sideOffset]="8"
                         [onOverlayEscapeKeyDownDisabled]="onOverlayEscapeKeyDownDisabled()"
                         [onOverlayOutsideClickDisabled]="onOverlayOutsideClickDisabled()"
-                        (onOpen)="countOpenTrue(true)"
-                        (onClosed)="countOpenTrue(false)"
                         rdxTooltipContent
                     >
                         <div class="TooltipContent" rdxTooltipContentAttributes>
@@ -162,39 +135,14 @@ import { WithOptionPanelComponent } from './utils/with-option-panel.component';
         </tooltip-with-option-panel>
     `
 })
-export class RdxTooltipTriggeringComponent extends OptionPanelBase {
+export class RdxTooltipAnchorComponent extends OptionPanelBase {
     readonly rootDirective1 = viewChild('root1');
     readonly rootDirective2 = viewChild('root2');
 
     readonly MountainSnowIcon = MountainSnow;
     readonly XIcon = X;
-
-    isOpenFalse = signal(false);
-    counterOpenFalse = signal(0);
-    externalControlFalse = signal(true);
-
-    isOpenTrue = signal(true);
-    counterOpenTrue = signal(0);
-    externalControlTrue = signal(true);
-
-    triggerOpenFalse(): void {
-        this.isOpenFalse.update((value) => !value);
-    }
-
-    countOpenFalse(open: boolean): void {
-        this.isOpenFalse.set(open);
-        this.counterOpenFalse.update((value) => value + 1);
-    }
-
-    triggerOpenTrue(): void {
-        this.isOpenTrue.update((value) => !value);
-    }
-
-    countOpenTrue(open: boolean): void {
-        this.isOpenTrue.set(open);
-        this.counterOpenTrue.update((value) => value + 1);
-    }
-
-    protected readonly containerAlert = containerAlert;
-    protected readonly TriangleAlert = TriangleAlert;
+    readonly LucideMapPinPlusInside = MapPinPlus;
+    readonly LucideMapPinPlus = MapPin;
+    readonly TriangleAlert = TriangleAlert;
+    readonly containerAlert = containerAlert;
 }
