+++ packages/primitives/hover-card/stories/hover-card-animations.component.ts
@@ -2,8 +2,8 @@ import { Component, signal, viewChild } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { RdxPositionAlign, RdxPositionSide } from '@radix-ng/primitives/core';
 import { LucideAngularModule, MountainSnow, TriangleAlert, X } from 'lucide-angular';
-import { RdxTooltipModule, RdxTooltipRootDirective } from '../index';
-import { RdxTooltipContentAttributesComponent } from '../src/tooltip-content-attributes.component';
+import { RdxHoverCardModule, RdxTooltipRootDirective } from '../index';
+import { RdxTooltipContentAttributesComponent } from '../src/hover-card-content-attributes.component';
 import { provideRdxCdkEventService } from '../src/utils/cdk-event.service';
 import { containerAlert } from './utils/constants';
 import { OptionPanelBase } from './utils/option-panel-base.class';
@@ -11,16 +11,16 @@ import styles from './utils/styles.constants';
 import { WithOptionPanelComponent } from './utils/with-option-panel.component';
 
 @Component({
-    selector: 'rdx-tooltip-positioning',
+    selector: 'rdx-tooltip-animations',
     providers: [provideRdxCdkEventService()],
     imports: [
         FormsModule,
-        RdxTooltipModule,
+        RdxHoverCardModule,
         LucideAngularModule,
         RdxTooltipContentAttributesComponent,
         WithOptionPanelComponent
     ],
-    styles: styles(),
+    styles: styles(true),
     template: `
         <tooltip-with-option-panel
             [arrowWidth]="arrowWidth()"
@@ -35,33 +35,20 @@ import { WithOptionPanelComponent } from './utils/with-option-panel.component';
             (closeDelayChange)="closeDelay.set($event)"
         >
             <div class="ParamsContainer">
-                Side:
-                <select [ngModel]="selectedSide()" (ngModelChange)="selectedSide.set($event)">
-                    <option [value]="sides.Top">{{ sides.Top }}</option>
-                    <option [value]="sides.Bottom">{{ sides.Bottom }}</option>
-                    <option [value]="sides.Left">{{ sides.Left }}</option>
-                    <option [value]="sides.Right">{{ sides.Right }}</option>
-                </select>
-                Align:
-                <select [ngModel]="selectedAlign()" (ngModelChange)="selectedAlign.set($event)">
-                    <option [value]="aligns.Center">{{ aligns.Center }}</option>
-                    <option [value]="aligns.Start">{{ aligns.Start }}</option>
-                    <option [value]="aligns.End">{{ aligns.End }}</option>
-                </select>
-                SideOffset:
-                <input [ngModel]="sideOffset()" (ngModelChange)="sideOffset.set($event)" type="number" />
-                AlignOffset:
-                <input [ngModel]="alignOffset()" (ngModelChange)="alignOffset.set($event)" type="number" />
-            </div>
-
-            <div class="ParamsContainer">
+                <input [ngModel]="cssAnimation()" (ngModelChange)="cssAnimation.set($event)" type="checkbox" />
+                CSS Animation
                 <input
-                    [ngModel]="disableAlternatePositions()"
-                    (ngModelChange)="disableAlternatePositions.set($event)"
+                    [ngModel]="cssOpeningAnimation()"
+                    (ngModelChange)="cssOpeningAnimation.set($event)"
                     type="checkbox"
                 />
-                Disable alternate positions (to see the result, scroll the page to make the tooltip cross the viewport
-                boundary)
+                On Opening Animation
+                <input
+                    [ngModel]="cssClosingAnimation()"
+                    (ngModelChange)="cssClosingAnimation.set($event)"
+                    type="checkbox"
+                />
+                On Closing Animation
             </div>
 
             <div class="ContainerAlerts">
@@ -69,24 +56,26 @@ import { WithOptionPanelComponent } from './utils/with-option-panel.component';
                 {{ containerAlert }}
             </div>
             <div class="container">
-                <ng-container [openDelay]="openDelay()" [closeDelay]="closeDelay()" rdxTooltipRoot>
-                    <button class="reset IconButton" rdxTooltipTrigger>
+                <ng-container
+                    [cssAnimation]="cssAnimation()"
+                    [cssOpeningAnimation]="cssOpeningAnimation()"
+                    [cssClosingAnimation]="cssClosingAnimation()"
+                    [openDelay]="openDelay()"
+                    [closeDelay]="closeDelay()"
+                    rdxTooltipRoot
+                >
+                    <button class="IconButton reset" rdxTooltipTrigger>
                         <lucide-angular [img]="MountainSnowIcon" size="16" style="display: flex" />
                     </button>
 
                     <ng-template
-                        [sideOffset]="sideOffset()"
-                        [alignOffset]="alignOffset()"
-                        [side]="selectedSide()"
-                        [align]="selectedAlign()"
-                        [alternatePositionsDisabled]="disableAlternatePositions()"
                         [onOverlayEscapeKeyDownDisabled]="onOverlayEscapeKeyDownDisabled()"
                         [onOverlayOutsideClickDisabled]="onOverlayOutsideClickDisabled()"
                         rdxTooltipContent
                     >
                         <div class="TooltipContent" rdxTooltipContentAttributes>
-                            <button class="reset TooltipClose" rdxTooltipClose aria-label="Close">
-                                <lucide-angular [img]="XIcon" size="12" style="display: flex" />
+                            <button class="TooltipClose reset" rdxTooltipClose aria-label="Close">
+                                <lucide-angular [img]="XIcon" size="16" style="display: flex" />
                             </button>
                             Add to library
                             <div
@@ -103,20 +92,18 @@ import { WithOptionPanelComponent } from './utils/with-option-panel.component';
         </tooltip-with-option-panel>
     `
 })
-export class RdxTooltipPositioningComponent extends OptionPanelBase {
+export class RdxTooltipAnimationsComponent extends OptionPanelBase {
     readonly rootDirective = viewChild(RdxTooltipRootDirective);
 
-    readonly selectedSide = signal(RdxPositionSide.Top);
-    readonly selectedAlign = signal(RdxPositionAlign.Center);
-    readonly sideOffset = signal<number | undefined>(void 0);
-    readonly alignOffset = signal<number | undefined>(void 0);
-    readonly disableAlternatePositions = signal(false);
+    readonly MountainSnowIcon = MountainSnow;
+    readonly XIcon = X;
 
     readonly sides = RdxPositionSide;
     readonly aligns = RdxPositionAlign;
 
-    readonly MountainSnowIcon = MountainSnow;
-    readonly XIcon = X;
-    protected readonly containerAlert = containerAlert;
+    cssAnimation = signal<boolean>(true);
+    cssOpeningAnimation = signal(true);
+    cssClosingAnimation = signal(true);
     protected readonly TriangleAlert = TriangleAlert;
+    protected readonly containerAlert = containerAlert;
 }
