+++ packages/primitives/hover-card/stories/hover-card-events.components.ts
@@ -1,9 +1,9 @@
-import { Component, signal, viewChild } from '@angular/core';
+import { Component, viewChild } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { RdxPositionAlign, RdxPositionSide } from '@radix-ng/primitives/core';
 import { LucideAngularModule, MountainSnow, TriangleAlert, X } from 'lucide-angular';
-import { RdxTooltipModule, RdxTooltipRootDirective } from '../index';
-import { RdxTooltipContentAttributesComponent } from '../src/tooltip-content-attributes.component';
+import { RdxHoverCardModule, RdxTooltipRootDirective } from '../index';
+import { RdxTooltipContentAttributesComponent } from '../src/hover-card-content-attributes.component';
 import { provideRdxCdkEventService } from '../src/utils/cdk-event.service';
 import { containerAlert } from './utils/constants';
 import { OptionPanelBase } from './utils/option-panel-base.class';
@@ -11,16 +11,18 @@ import styles from './utils/styles.constants';
 import { WithOptionPanelComponent } from './utils/with-option-panel.component';
 
 @Component({
-    selector: 'rdx-tooltip-animations',
+    selector: 'rdx-tooltip-events',
     providers: [provideRdxCdkEventService()],
     imports: [
-        FormsModule,
-        RdxTooltipModule,
+        RdxHoverCardModule,
         LucideAngularModule,
+        FormsModule,
         RdxTooltipContentAttributesComponent,
         WithOptionPanelComponent
     ],
-    styles: styles(true),
+    styles: `
+        ${styles()}
+    `,
     template: `
         <tooltip-with-option-panel
             [arrowWidth]="arrowWidth()"
@@ -34,48 +36,25 @@ import { WithOptionPanelComponent } from './utils/with-option-panel.component';
             (openDelayChange)="openDelay.set($event)"
             (closeDelayChange)="closeDelay.set($event)"
         >
-            <div class="ParamsContainer">
-                <input [ngModel]="cssAnimation()" (ngModelChange)="cssAnimation.set($event)" type="checkbox" />
-                CSS Animation
-                <input
-                    [ngModel]="cssOpeningAnimation()"
-                    (ngModelChange)="cssOpeningAnimation.set($event)"
-                    type="checkbox"
-                />
-                On Opening Animation
-                <input
-                    [ngModel]="cssClosingAnimation()"
-                    (ngModelChange)="cssClosingAnimation.set($event)"
-                    type="checkbox"
-                />
-                On Closing Animation
-            </div>
-
             <div class="ContainerAlerts">
                 <lucide-angular [img]="TriangleAlert" size="16" />
                 {{ containerAlert }}
             </div>
             <div class="container">
-                <ng-container
-                    [cssAnimation]="cssAnimation()"
-                    [cssOpeningAnimation]="cssOpeningAnimation()"
-                    [cssClosingAnimation]="cssClosingAnimation()"
-                    [openDelay]="openDelay()"
-                    [closeDelay]="closeDelay()"
-                    rdxTooltipRoot
-                >
-                    <button class="IconButton reset" rdxTooltipTrigger>
+                <ng-container [openDelay]="openDelay()" [closeDelay]="closeDelay()" rdxTooltipRoot>
+                    <button class="reset IconButton" #triggerElement rdxTooltipTrigger>
                         <lucide-angular [img]="MountainSnowIcon" size="16" style="display: flex" />
                     </button>
 
                     <ng-template
                         [onOverlayEscapeKeyDownDisabled]="onOverlayEscapeKeyDownDisabled()"
                         [onOverlayOutsideClickDisabled]="onOverlayOutsideClickDisabled()"
+                        [sideOffset]="8"
                         rdxTooltipContent
                     >
                         <div class="TooltipContent" rdxTooltipContentAttributes>
-                            <button class="TooltipClose reset" rdxTooltipClose aria-label="Close">
-                                <lucide-angular [img]="XIcon" size="16" style="display: flex" />
+                            <button class="reset TooltipClose" rdxTooltipClose aria-label="Close">
+                                <lucide-angular [img]="XIcon" size="12" style="display: flex" />
                             </button>
                             Add to library
                             <div
@@ -92,18 +71,14 @@ import { WithOptionPanelComponent } from './utils/with-option-panel.component';
         </tooltip-with-option-panel>
     `
 })
-export class RdxTooltipAnimationsComponent extends OptionPanelBase {
+export class RdxTooltipEventsComponent extends OptionPanelBase {
     readonly rootDirective = viewChild(RdxTooltipRootDirective);
 
     readonly MountainSnowIcon = MountainSnow;
     readonly XIcon = X;
 
-    readonly sides = RdxPositionSide;
-    readonly aligns = RdxPositionAlign;
-
-    cssAnimation = signal<boolean>(true);
-    cssOpeningAnimation = signal(true);
-    cssClosingAnimation = signal(true);
-    protected readonly TriangleAlert = TriangleAlert;
+    protected readonly sides = RdxPositionSide;
+    protected readonly aligns = RdxPositionAlign;
     protected readonly containerAlert = containerAlert;
+    protected readonly TriangleAlert = TriangleAlert;
 }
