+++ packages/primitives/hover-card/stories/hover-card-triggering.component.ts
@@ -1,9 +1,8 @@
-import { Component, viewChild } from '@angular/core';
+import { Component, signal, viewChild } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { LucideAngularModule, MapPin, MapPinPlus, MountainSnow, TriangleAlert, X } from 'lucide-angular';
-import { RdxTooltipModule } from '../index';
-import { RdxTooltipAnchorDirective } from '../src/tooltip-anchor.directive';
-import { RdxTooltipContentAttributesComponent } from '../src/tooltip-content-attributes.component';
+import { LucideAngularModule, MountainSnow, TriangleAlert, X } from 'lucide-angular';
+import { RdxHoverCardModule } from '../index';
+import { RdxTooltipContentAttributesComponent } from '../src/hover-card-content-attributes.component';
 import { provideRdxCdkEventService } from '../src/utils/cdk-event.service';
 import { containerAlert } from './utils/constants';
 import { OptionPanelBase } from './utils/option-panel-base.class';
@@ -11,19 +10,18 @@ import styles from './utils/styles.constants';
 import { WithOptionPanelComponent } from './utils/with-option-panel.component';
 
 @Component({
-    selector: 'rdx-tooltip-anchor',
+    selector: 'rdx-tooltip-triggering',
     providers: [provideRdxCdkEventService()],
     imports: [
         FormsModule,
-        RdxTooltipModule,
+        RdxHoverCardModule,
         LucideAngularModule,
         RdxTooltipContentAttributesComponent,
-        WithOptionPanelComponent,
-        RdxTooltipAnchorDirective
+        WithOptionPanelComponent
     ],
     styles: styles(),
     template: `
-        <p class="ExampleSubtitle">Internal Anchor (within TooltipRoot)</p>
+        <p class="ExampleSubtitle">Initially closed</p>
         <tooltip-with-option-panel
             [arrowWidth]="arrowWidth()"
             [arrowHeight]="arrowHeight()"
@@ -36,6 +34,20 @@ import { WithOptionPanelComponent } from './utils/with-option-panel.component';
             (openDelayChange)="openDelay.set($event)"
             (closeDelayChange)="closeDelay.set($event)"
         >
+            <div class="ParamsContainer">
+                <button (mouseup)="triggerOpenFalse()" type="button">Open: {{ isOpenFalse() }}</button>
+                onOpenChange count: {{ counterOpenFalse() }}
+            </div>
+
+            <div class="ParamsContainer">
+                <input
+                    [ngModel]="externalControlFalse()"
+                    (ngModelChange)="externalControlFalse.set($event)"
+                    type="checkbox"
+                />
+                External control
+            </div>
+
             <div class="ContainerAlerts">
                 <lucide-angular [img]="TriangleAlert" size="16" />
                 {{ containerAlert }}
@@ -43,21 +55,22 @@ import { WithOptionPanelComponent } from './utils/with-option-panel.component';
             <div class="container">
                 <ng-container
                     #root1="rdxTooltipRoot"
+                    [open]="isOpenFalse()"
                     [openDelay]="openDelay()"
                     [closeDelay]="closeDelay()"
+                    [externalControl]="externalControlFalse()"
                     rdxTooltipRoot
                 >
-                    <button class="reset IconButton InternalAnchor" rdxTooltipAnchor>
-                        <lucide-angular [img]="LucideMapPinPlusInside" size="16" style="display: flex" />
-                    </button>
-
                     <button class="reset IconButton" rdxTooltipTrigger>
                         <lucide-angular [img]="MountainSnowIcon" size="16" style="display: flex" />
                     </button>
 
                     <ng-template
+                        [sideOffset]="8"
                         [onOverlayEscapeKeyDownDisabled]="onOverlayEscapeKeyDownDisabled()"
                         [onOverlayOutsideClickDisabled]="onOverlayOutsideClickDisabled()"
+                        (onOpen)="countOpenFalse(true)"
+                        (onClosed)="countOpenFalse(false)"
                         rdxTooltipContent
                     >
                         <div class="TooltipContent" rdxTooltipContentAttributes>
@@ -78,7 +91,7 @@ import { WithOptionPanelComponent } from './utils/with-option-panel.component';
             <div class="TooltipId">ID: {{ rootDirective1()?.uniqueId() }}</div>
         </tooltip-with-option-panel>
 
-        <p class="ExampleSubtitle">External Anchor (outside TooltipRoot)</p>
+        <p class="ExampleSubtitle">Initially open</p>
         <tooltip-with-option-panel
             [arrowWidth]="arrowWidth()"
             [arrowHeight]="arrowHeight()"
@@ -91,20 +104,31 @@ import { WithOptionPanelComponent } from './utils/with-option-panel.component';
             (openDelayChange)="openDelay.set($event)"
             (closeDelayChange)="closeDelay.set($event)"
         >
+            <div class="ParamsContainer">
+                <button (mouseup)="triggerOpenTrue()" type="button">Open: {{ isOpenTrue() }}</button>
+                <span>onOpenChange count: {{ counterOpenTrue() }}</span>
+            </div>
+
+            <div class="ParamsContainer">
+                <input
+                    [ngModel]="externalControlTrue()"
+                    (ngModelChange)="externalControlTrue.set($event)"
+                    type="checkbox"
+                />
+                External control
+            </div>
+
             <div class="ContainerAlerts">
                 <lucide-angular [img]="TriangleAlert" size="16" />
                 {{ containerAlert }}
             </div>
             <div class="container">
-                <button class="reset IconButton ExternalAnchor" #externalAnchor="rdxTooltipAnchor" rdxTooltipAnchor>
-                    <lucide-angular [img]="LucideMapPinPlus" size="16" style="display: flex" />
-                </button>
-
                 <ng-container
                     #root2="rdxTooltipRoot"
-                    [anchor]="externalAnchor"
+                    [open]="isOpenTrue()"
                     [openDelay]="openDelay()"
                     [closeDelay]="closeDelay()"
+                    [externalControl]="externalControlTrue()"
                     rdxTooltipRoot
                 >
                     <button class="reset IconButton" rdxTooltipTrigger>
@@ -112,8 +136,11 @@ import { WithOptionPanelComponent } from './utils/with-option-panel.component';
                     </button>
 
                     <ng-template
+                        [sideOffset]="8"
                         [onOverlayEscapeKeyDownDisabled]="onOverlayEscapeKeyDownDisabled()"
                         [onOverlayOutsideClickDisabled]="onOverlayOutsideClickDisabled()"
+                        (onOpen)="countOpenTrue(true)"
+                        (onClosed)="countOpenTrue(false)"
                         rdxTooltipContent
                     >
                         <div class="TooltipContent" rdxTooltipContentAttributes>
@@ -135,14 +162,39 @@ import { WithOptionPanelComponent } from './utils/with-option-panel.component';
         </tooltip-with-option-panel>
     `
 })
-export class RdxTooltipAnchorComponent extends OptionPanelBase {
+export class RdxTooltipTriggeringComponent extends OptionPanelBase {
     readonly rootDirective1 = viewChild('root1');
     readonly rootDirective2 = viewChild('root2');
 
     readonly MountainSnowIcon = MountainSnow;
     readonly XIcon = X;
-    readonly LucideMapPinPlusInside = MapPinPlus;
-    readonly LucideMapPinPlus = MapPin;
-    readonly TriangleAlert = TriangleAlert;
-    readonly containerAlert = containerAlert;
+
+    isOpenFalse = signal(false);
+    counterOpenFalse = signal(0);
+    externalControlFalse = signal(true);
+
+    isOpenTrue = signal(true);
+    counterOpenTrue = signal(0);
+    externalControlTrue = signal(true);
+
+    triggerOpenFalse(): void {
+        this.isOpenFalse.update((value) => !value);
+    }
+
+    countOpenFalse(open: boolean): void {
+        this.isOpenFalse.set(open);
+        this.counterOpenFalse.update((value) => value + 1);
+    }
+
+    triggerOpenTrue(): void {
+        this.isOpenTrue.update((value) => !value);
+    }
+
+    countOpenTrue(open: boolean): void {
+        this.isOpenTrue.set(open);
+        this.counterOpenTrue.update((value) => value + 1);
+    }
+
+    protected readonly containerAlert = containerAlert;
+    protected readonly TriangleAlert = TriangleAlert;
 }
