+++ packages/primitives/hover-card/stories/hover-card.docs.mdx
@@ -1,13 +1,13 @@
 import {ArgTypes, Canvas, Markdown, Meta, Source} from '@storybook/blocks';
-import * as TooltipStories from "./tooltip.stories";
-import {RdxTooltipRootDirective} from "../src/tooltip-root.directive";
-import {RdxTooltipContentDirective} from "../src/tooltip-content.directive";
-import {RdxTooltipArrowDirective} from "../src/tooltip-arrow.directive";
-import {RdxTooltipAnchorDirective} from "../src/tooltip-anchor.directive";
-import {RdxTooltipTriggerDirective} from "../src/tooltip-trigger.directive";
-import {RdxTooltipCloseDirective} from "../src/tooltip-close.directive";
+import * as TooltipStories from "./hover-card.stories";
+import {RdxTooltipRootDirective} from "../src/hover-card-root.directive";
+import {RdxTooltipContentDirective} from "../src/hover-card-content.directive";
+import {RdxTooltipArrowDirective} from "../src/hover-card-arrow.directive";
+import {RdxTooltipAnchorDirective} from "../src/hover-card-anchor.directive";
+import {RdxTooltipTriggerDirective} from "../src/hover-card-trigger.directive";
+import {RdxTooltipCloseDirective} from "../src/hover-card-close.directive";
 import {animationStylesOnly} from "./utils/styles.constants";
-import {RdxTooltipContentAttributesComponent} from "../src/tooltip-content-attributes.component";
+import {RdxTooltipContentAttributesComponent} from "../src/hover-card-content-attributes.component";
 
 <Meta title="Primitives/Tooltip" />
