+++ packages/primitives/hover-card/stories/utils/containers.registry.ts
@@ -1,5 +1,5 @@
 import { isDevMode } from '@angular/core';
-import { RdxTooltipRootDirective } from '../../src/tooltip-root.directive';
+import { RdxTooltipRootDirective } from '../../src/hover-card-root.directive';
 import { injectRdxCdkEventService } from '../../src/utils/cdk-event.service';
 
 const containerRegistry: Map<HTMLElement, RdxTooltipRootDirective> = new Map();
