+++ packages/primitives/hover-card/stories/utils/option-panel-base.class.ts
@@ -1,6 +1,6 @@
 import { afterNextRender, DestroyRef, Directive, ElementRef, inject, signal, viewChildren } from '@angular/core';
 import { injectDocument, RDX_POSITIONING_DEFAULTS } from '@radix-ng/primitives/core';
-import { RdxTooltipRootDirective } from '../../src/tooltip-root.directive';
+import { RdxTooltipRootDirective } from '../../src/hover-card-root.directive';
 import { injectRdxCdkEventService } from '../../src/utils/cdk-event.service';
 import { deregisterContainer, registerContainer, setRdxCdkEventService } from './containers.registry';
 import { IArrowDimensions, IIgnoreClickOutsideContainer, IOpenCloseDelay } from './types';
